New QueueTees queue
The QueueTees class creates a queue list using the EnQueue() method to add objects to the queue and the deQueue() method to remove objects from the queue based on whichever object was added to the queue most recently. (Sorry for turning this in late, I swear I finished it on time I just forgot to push it until now).